Move Dagger annotations to common
