+++ kirinmaru-android/src/main/java/commons/android/dagger/Qualifiers.kt
@@ -1,0 +1,11 @@
+package commons.android.dagger
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ApplicationContext
+
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ActivityContext
