+++ kirinmaru-android/src/main/java/commons/android/dagger/Scopes.kt
@@ -1,0 +1,23 @@
+package commons.android.dagger
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ApplicationScope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ActivityScope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ServiceScope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class FragmentScope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class ViewModelScope
