--- kirinmaru-android/src/main/java/stream/reconfig/kirinmaru/android/di/qualifiers/ActivityContext.kt
@@ -1,10 +1,0 @@
-package stream.reconfig.kirinmaru.android.di.qualifiers
-
-import javax.inject.Qualifier
-
-/**
- *
- */
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ActivityContext
