--- kirinmaru-android/src/main/java/stream/reconfig/kirinmaru/android/di/scopes/ApplicationScope.kt
@@ -1,7 +1,0 @@
-package stream.reconfig.kirinmaru.android.di.scopes
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ApplicationScope
